Revert "fix go formatting of mlx runner"
This reverts commit 4eec06645170ac1654ac4ecd1f83d8707e68c81c.

+++ x/imagegen/models/zimage/vae.go
@@ -135,7 +135,7 @@ func (gn *GroupNormLayer) forwardTiled(x *mlx.Array, B, H, W, C int32) *mlx.Arra
 	mlx.Eval(mean, invStd)
 
 	// Tile along H dimension
-	tileH := 512 * 512 / W
+	tileH := int32(512 * 512 / W)
 	if tileH < 1 {
 		tileH = 1
 	}
